requirements.txt
asgiref==3.5.2
Django==4.1.4
django-froala-editor==4.0.16
django-js-asset==2.0.0
Pillow==9.3.0
sqlparse==0.4.3
tzdata==2022.7
